+++ examples/hello_raylib.rs
@@ -1,16 +1,17 @@
+use raylib::core::game_loop;
 use raylib::prelude::*;
 
 fn main() {
-    let (mut rl, thread) = raylib::init()
+    let (rl, thread) = raylib::init()
         .size(640, 480)
         .title("Hello, Raylib")
         .highdpi()
         .build();
 
-    while !rl.window_should_close() {
-        let mut d = rl.begin_drawing(&thread);
+    game_loop::run(rl, thread, 60, |rl, thread| {
+        let mut d = rl.begin_drawing(thread);
 
         d.clear_background(Color::WHITE);
         d.draw_text("Hello, Raylib", 12, 12, 20, Color::BLACK);
-    }
+    });
 }